Added global error handling test
