+++ __tests__/supertest.js
@@ -8,11 +8,8 @@ const db = require('../server/models/model');
  * Read the docs! https://www.npmjs.com/package/supertest
  */
 
-/**
- * Read the docs! https://www.npmjs.com/package/supertest
- */
 describe('Route integration', () => {
-  describe('TESTING ALL', () => {
+  describe('/test', () => {
     describe('GET', () => {
       // Note that we return the evaluation of `request` here! It evaluates to
       // a promise, so Jest knows not to say this test passes until that
@@ -43,5 +40,17 @@ describe('Route integration', () => {
 
       xit('responds to invalid request with 400 status and error message in body', () => {});
     });
+  });
+
+  describe('Global Error Handler', () => {
+    it('returns a particular message and 404 status', async () => {
+      const res = await request(server).post('/video');
+      expect(res.status).toBe(400);
+      expect(res.body.message).toBe('file not sent by the client');
+    });
+    // return request(server)
+    //     .get('/test')
+    //     .expect('Content-Type', /json/)
+    //     .expect(200);
   });
 });
